requirements.txt
Django==4.0.6
django-autoslug==1.9.8
Pillow==9.2.0
django-phonenumber-field==6.3.0